Sync policy cache in Hive back to Atlas
If the Atlas DB is dropped for some reason Atlas and Ranger will come out of sync.
This commit add functionality to use a policy cache file copied from Hive and
write that state into Atlas. This way Atlas and Ranger will comback to a common
picture of the state.
After the command has runned it is recommended to apply your own rules agin with
policytool.